Refactor test_countries
